Refactor Marathon task details into own component
